Fix format checking
